pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "outgoing"
dynamic = ["version"]
description = "Common interface for multiple e-mail methods"
readme = "README.rst"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "John Thorvald Wodder II", email = "outgoing@varonathe.org" }
]

keywords = [
    "e-mail",
    "email",
    "mailbox",
    "mbox",
    "send mail",
    "sendmail",
    "smtp",
]

classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: Communications :: Email",
    "Typing :: Typed",
]

dependencies = [
    "keyring        >= 21.7",
    "morecontext    ~= 0.4",
    "platformdirs   >= 3.0, < 5.0",
    "pydantic       ~= 2.0",
    "python-dotenv  >= 0.11, < 2.0",
    "tomli          >= 1.2, < 3.0; python_version < '3.11'",
]

[project.scripts]
outgoing = "outgoing.__main__:main"

[project.entry-points."outgoing.senders"]
babyl = "outgoing.senders.mailboxes:BabylSender"
command = "outgoing.senders.command:CommandSender"
maildir = "outgoing.senders.mailboxes:MaildirSender"
mbox = "outgoing.senders.mailboxes:MboxSender"
mh = "outgoing.senders.mailboxes:MHSender"
mmdf = "outgoing.senders.mailboxes:MMDFSender"
null = "outgoing.senders.null:NullSender"
smtp = "outgoing.senders.smtp:SMTPSender"

[project.entry-points."outgoing.password_schemes"]
dotenv = "outgoing.passwords:dotenv_scheme"
base64 = "outgoing.passwords:base64_scheme"
env = "outgoing.passwords:env_scheme"
file = "outgoing.passwords:file_scheme"
keyring = "outgoing.passwords:keyring_scheme"

[project.urls]
"Source Code" = "https://github.com/jwodder/outgoing"
"Bug Tracker" = "https://github.com/jwodder/outgoing/issues"
"Documentation" = "https://outgoing.readthedocs.io"

[tool.hatch.version]
path = "src/outgoing/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/docs",
    "/src",
    "/test",
    "CHANGELOG.*",
    "CONTRIBUTORS.*",
    "tox.ini",
]

[tool.mypy]
allow_incomplete_defs = false
allow_untyped_defs = false
ignore_missing_imports = false
# <https://github.com/python/mypy/issues/7773>:
no_implicit_optional = true
implicit_reexport = false
local_partial_types = true
pretty = true
show_error_codes = true
show_traceback = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
plugins = ["pydantic.mypy"]
exclude = "test/data/"

[tool.pydantic-mypy]
init_forbid_extra = true
warn_required_dynamic_aliases = true
